+++ ssaUtils/testdata/NestedFunctionWithLock/NestedFunctionWithLock.go
