+++ ssaUtils/testdata/NestedFunctionWithLockButThenUnlock/NestedFunctionWithLockButThenUnlock.go
