+++ ssaUtils/testdata/RecursionNestedFunctionWithLock/RecursionNestedFunctionWithLock.go
