+++ ssaUtils/testdata/functionWithEvenLockAndDeferUnlock/functionWithEvenLockAndDeferUnlock.go
