+++ ssaUtils/testdata/functionWithEvenLockAndUnlock/functionWithEvenLockAndUnlock.go
