+++ ssaUtils/testdata/functionWithLock/functionWithLock.go
