+++ ssaUtils/testdata/functionWithoutLocks/functionWithoutLocks.go
